// src/types/models/Customer.ts
export type Customer = {
  id: string;
  customerNumber?: string | undefined;
  email?: string | undefined;
  houseNumber?: string | undefined;
  firstname?: string | undefined;
  city?: string | undefined;
  postalCode?: string | undefined;
  street?: string | undefined;
  phoneNumber?: string | undefined;
  lastname?: string | undefined;
};